Add support for reading from s3 zarr tiles
